# api/openapi.yaml
openapi: 3.0.3
info:
  title: Tarot App API
  version: 0.1.0
servers:
  - url: https://api.example.com
security:
  - bearerAuth: []
paths:
...
  /billing/verify/purchase:
    post:
      summary: Verify one-time purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseVerifyRequest'
      responses:
        '200':
          description: Verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /billing/verify/subscription:
    post:
      summary: Verify subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionVerifyRequest'
      responses:
        '200':
          description: Verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /billing/status:
    get:
      summary: Billing status
      responses:
        '200':
          description: Billing status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Too Many Requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    PurchaseVerifyRequest:
      type: object
      required: [platform, receipt]
      properties:
        platform:
          type: string
          enum: [android]
        receipt:
          type: string
        product_key:
          type: string
    SubscriptionVerifyRequest:
      type: object
      required: [platform, receipt]
      properties:
        platform:
          type: string
          enum: [android]
        receipt:
          type: string
    BillingStatus:
      type: object
      properties:
        subscription_active:
          type: boolean
        subscription_expires_at:
          type: string
          format: date-time
          nullable: true
        ads_disabled:
          type: boolean
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean